Added the ability to lose the game.
There are two scenes, Win and Lose. You can get to them by sending a
message to the endDecider with the function either Win or Lose.

Upon entering these scenes, it will display a Win or Lose screen for 3
mandatory seconds, after which any input will cause the game to restart.